internal: document usage of structToQuery and reflection
